Move getConnectionsFromConfig out of connections model
